HRP/HOUSE RENT PREDICTION/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
flask-bcrypt
folium
python-dotenv